requirements.txt
selene==2.0.0rc8
pytest~=7.4.3
allure-pytest
python-dotenv
pydantic-settings
requests
jsonschema
curlify